requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Django==4.1.6
django-redis==5.2.0
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
redis==4.5.1
sqlparse==0.4.3
tzdata==2022.7
